# Faster streaming 'jsonl' parser