--- a/extensions/ql-vscode/test/benchmarks/jsonl-reader.bench.ts
+++ b/extensions/ql-vscode/test/benchmarks/jsonl-reader.bench.ts
@@ -1,9 +1,7 @@
 import { readFile } from "fs-extra";
-import { readJsonlFile, readJsonlFile2 } from "../../src/common/jsonl-reader";
+import { readJsonlFile } from "../../src/common/jsonl-reader";
 import { performance } from "perf_hooks";
 import { join } from "path";
-import { createReadStream } from "fs";
-import { createInterface } from "readline";
 
 /** An "obviously correct" implementation to test against. */
 async function readJsonlReferenceImpl<T>(
@@ -21,18 +19,6 @@ async function readJsonlReferenceImpl<T>(
   }
 }
 
-async function justReadline(
-  path: string,
-  handler: (value: unknown) => Promise<void>,
-) {
-  const stream = createReadStream(path, "utf8");
-  const rl = createInterface(stream);
-
-  for await (const line of rl) {
-    await handler(line);
-  }
-}
-
 type ParserFn = (
   text: string,
   callback: (v: unknown) => Promise<void>,
@@ -41,8 +27,6 @@ type ParserFn = (
 const parsers: Record<string, ParserFn> = {
   readJsonlReferenceImpl,
   readJsonlFile,
-  readJsonlFile2,
-  justReadline,
 };
 
 async function main() {