Replaces the streaming JSONL parser with a faster one that doesn't call `readline`. The streaming JSONL parser on our branch was slower than the original sync version (but doesn't run out of memory), and it seems `readline` is a bottleneck:

Running the benchmark on a 21 MB logfile:
- `readJsonlReferenceImpl`: 172.4 ms (original non-streaming version)
- `readJsonlFile`: 283.3 ms (streaming version based on `readline`)
- `readJsonlFile2`: 151.3 ms (🚀  new version without `readline`)
- `justReadline`: 187.5 ms (consumes the file with `readline` and nothing else)

Note: targeting the hackathon branch. I think it's best to polish that up and try to merge it into `main`.