@esbena 2024-11-18
nice! I would expect the non-streaming version to be faster for small files like the one you benchmarked on, and would have advocated that as the default for small files. 
For completeness, can you try on some larger files like https://github.com/codeql-dca-runners/codeql-dca-worker_javascript/actions/runs/11840886420/artifacts/2188663427 (amphtml on https://github.com/github/codeql-dca-main/issues/24803)