@asgerf 2024-11-18
I ran it anyway:
- `readJsonlReferenceImpl`: out of memory
- `readJsonlFile`: 6439.4 ms
- `readJsonlFile2`: 3538.4 ms
- `justReadline`: 3664.3 ms