@esbena 2024-11-18
My two guesses:
- `readline` reads smaller chunks than a plain `readStream` does by default
- `readline` [has many string checking operations per chunk](https://github.com/nodejs/node/blob/64c647e3421e7756d68b4b12ee45e64d101b2956/lib/internal/readline/interface.js#L575-L620), the `readStream` implementation only cares for `\n`