@asgerf 2024-11-19
Although nobody formally 'Approved' the PR I'm gonna merge to the hackathon branch so I can start preparing the PR against main.